+++ common/src/main/java/com/faboslav/friendsandfoes/FriendsAndFoes.java
@@ -11,8 +11,8 @@ import org.slf4j.LoggerFactory;
 
 public final class FriendsAndFoes
 {
-	private static final FriendsAndFoesConfig CONFIG = OmegaConfig.register(FriendsAndFoesConfig.class);
 	private static final Logger LOGGER = LoggerFactory.getLogger(FriendsAndFoes.MOD_ID);
+	private static final FriendsAndFoesConfig CONFIG = OmegaConfig.register(FriendsAndFoesConfig.class);
 	public static final String MOD_ID = "friendsandfoes";
 	public static final ServerTickDeltaCounter serverTickDeltaCounter = new ServerTickDeltaCounter(20.0F, 0L);
 
@@ -42,6 +42,7 @@ public final class FriendsAndFoes
 		FriendsAndFoesEntityTypes.init();
 		FriendsAndFoesItems.init();
 		FriendsAndFoesMemoryModuleTypes.init();
+		FriendsAndFoesParticleTypes.init();
 		FriendsAndFoesPointOfInterestTypes.init();
 		FriendsAndFoesSoundEvents.init();
 		FriendsAndFoesStructureProcessorTypes.init();