+++ common/src/main/java/com/faboslav/friendsandfoes/mixin/IllusionerEntityMixin.java
@@ -33,6 +33,9 @@ import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(IllusionerEntity.class)
 @SuppressWarnings({"rawtypes", "unchecked"})
@@ -69,18 +72,17 @@ public abstract class IllusionerEntityMixin extends SpellcastingIllagerEntity im
 		this.illusioner = null;
 	}
 
-	@Override
-	public void initDataTracker() {
-		super.initDataTracker();
-
+	@Inject(
+		at = @At("TAIL"),
+		method = "initDataTracker"
+	)
+	public void friendsandfoes_initDataTracker(CallbackInfo ci) {
 		if (FriendsAndFoes.getConfig().enableIllusioner == false) {
-			return;
+			this.dataTracker.startTracking(IS_ILLUSION, false);
+			this.dataTracker.startTracking(WAS_ATTACKED, false);
+			this.dataTracker.startTracking(TICKS_UNTIL_DESPAWN, 0);
+			this.dataTracker.startTracking(TICKS_UNTIL_CAN_CREATE_ILLUSIONS, 0);
 		}
-
-		this.dataTracker.startTracking(IS_ILLUSION, false);
-		this.dataTracker.startTracking(WAS_ATTACKED, false);
-		this.dataTracker.startTracking(TICKS_UNTIL_DESPAWN, 0);
-		this.dataTracker.startTracking(TICKS_UNTIL_CAN_CREATE_ILLUSIONS, 0);
 	}
 
 	@Override
@@ -286,8 +288,6 @@ public abstract class IllusionerEntityMixin extends SpellcastingIllagerEntity im
 			int y = (int) illusionerPosition.getY();
 			int z = (int) (illusionerPosition.getZ() + radius * MathHelper.sin(angle));
 
-			this.createIllusion(x, y, z);
-
 			if (randomPoint == point) {
 				boolean teleportResult = this.tryToTeleport(x, y, z);
 
@@ -295,6 +295,8 @@ public abstract class IllusionerEntityMixin extends SpellcastingIllagerEntity im
 					this.addStatusEffect(new StatusEffectInstance(StatusEffects.INVISIBILITY, INVISIBILITY_TICKS));
 					this.spawnCloudParticles();
 				}
+			} else {
+				this.createIllusion(x, y, z);
 			}
 		}
 	}