+++ common/src/main/java/com/faboslav/friendsandfoes/platform/RegistryHelper.java
@@ -14,6 +14,7 @@ import net.minecraft.entity.ai.brain.Activity;
 import net.minecraft.entity.ai.brain.MemoryModuleType;
 import net.minecraft.entity.attribute.DefaultAttributeContainer;
 import net.minecraft.item.Item;
+import net.minecraft.particle.DefaultParticleType;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.structure.processor.StructureProcessor;
 import net.minecraft.structure.processor.StructureProcessorType;
@@ -76,6 +77,11 @@ public final class RegistryHelper
 		String name,
 		Supplier<T> memoryModuleType
 	) {
+		throw new AssertionError();
+	}
+
+	@ExpectPlatform
+	public static void registerParticleType(String name, DefaultParticleType particleType) {
 		throw new AssertionError();
 	}